// src/app/services/openapi-client/model/createCategoryDto.ts
/**
 * ShopErpApi | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * DTO for creating a category
 */
export interface CreateCategoryDto { 
    parentId?: number | null;
    name: string;
    slug: string;
    description?: string | null;
    imageUrl?: string | null;
    displayOrder?: number;
    isActive?: boolean;
    metaTitle?: string | null;
    metaDescription?: string | null;
}